+++ relm4/src/extensions/set_child.rs
@@ -57,6 +57,7 @@ mod libadwaita {
         adw::Clamp,
         adw::ClampScrollable,
         adw::SplitButton,
-        adw::StatusPage
+        adw::StatusPage,
+        adw::ToastOverlay
     );
 }