+++ relm4/src/lib.rs
@@ -33,27 +33,24 @@ pub mod shutdown;
 /// Shared state that can be accessed by many components.
 pub mod shared_state;
 
-pub mod util;
-
 pub(crate) mod late_initialization;
 /// A simpler version of components that does work
 /// in the background.
 pub mod worker;
 
-pub use self::channel::{channel, Receiver, Sender};
-pub use self::component::{
+pub use channel::{channel, Receiver, Sender};
+pub use component::{
     Component, ComponentBuilder, ComponentController, ComponentParts, Controller, MessageBroker,
     OnDestroy, SimpleComponent,
 };
-pub use self::extensions::*;
-pub use self::sender::ComponentSender;
-pub use self::shared_state::SharedState;
-pub use self::shutdown::ShutdownReceiver;
-pub use self::worker::{Worker, WorkerController, WorkerHandle};
+pub use extensions::*;
+pub use sender::ComponentSender;
+pub use shared_state::SharedState;
+pub use shutdown::ShutdownReceiver;
+pub use worker::{Worker, WorkerController, WorkerHandle};
 
 pub use app::RelmApp;
 pub use tokio::task::JoinHandle;
-pub use util::{WidgetPlus, WidgetRef};
 
 use gtk::prelude::*;
 use once_cell::sync::OnceCell;